blktests: replace module removal with patient module removal
A long time ago, in a galaxy far, far away...

I ran into some odd scsi_debug false positives with fstests. This
prompted me to look into them given these false positives prevents
me from moving forward with establishing a test baseline with high
number of cycles. That is, this stupid issue was prevening creating
high confidence in testing.

I reported it [0] and exchanged some ideas with Doug. However, in
the end, despite efforts to help things with scsi_debug there were
still issues lingering which seemed to defy our expectations upstream.
One of the last hanging fruit issues is and always has been that
userspace expectations for proper module removal has been broken.

Long ago a WAIT option was added... then removed as it was deemed
not needed as folks couldn't figure out when these races happened.
The races are actually pretty easy to trigger, it was just never
properly documented. A simpe blkdev_open() will easily bump a module
refcnt, and these days many thing scan do that sort of thing.

The proper solution is to implement then a patient module removal
on kmod and patches have been sent for that and those patches are
under review. In the meantime we need a work around to open code a
similar solution for users of old versions of kmod. I sent an open
coded solution for fstests about 1 month ago. Now that that stuff
is merged and tested with more exposure, its time to match with
parity on blktests.

[0] https://bugzilla.kernel.org/show_bug.cgi?id=212337

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>